code: turned gameelement into stageelement
